Move parsing as a top level concern of Spoom
